requirements.txt
streamlit
google-genai==1.7.0
mysql-connector-python
pandas 
matplotlib 
seaborn
scipy 
pydantic-settings